Restyler.App.Class

Documentation

class (Monad m, MonadLogger m, MonadReader App m, MonadError AppError m) => MonadApp m where

Minimal complete definition

runGitHub, getCurrentDirectory, setCurrentDirectory, doesFileExist, readFile, exitSuccess, callProcess, readProcess, downloadFile

Methods

runGitHub :: Request k a -> m a

runGitHubFirst :: (FetchCount -> Request k (Vector a)) -> m (Maybe a)

Fetch the first page using runGitHub, return the first item

runGitHub_ :: Request k a -> m ()

void . runGitHub

getCurrentDirectory :: m FilePath

setCurrentDirectory :: FilePath -> m ()

doesFileExist :: FilePath -> m Bool

readFile :: FilePath -> m Text

exitSuccess :: m ()

callProcess :: String -> [String] -> m ()

readProcess :: String -> [String] -> String -> m String

downloadFile :: Text -> FilePath -> m ()

Re-exports

This is useful because it defines what GitHub APIs we use and takes care of the hidings we'd otherwise need to do at every import site.

module GitHub.Endpoints.Issues.Comments

module GitHub.Endpoints.PullRequests

module GitHub.Endpoints.Repos.Statuses