Restyler.App.Run

Documentation

bootstrapApp :: Options -> FilePath -> ExceptT AppError IO App

Bootstrap the initial App type

We want to have the PullRequest and Config in our application environment, so we need them to construct an App. However, it's valuable to use our normal AppM actions to interact with GitHub or call processes so we get the same logging and error-handling there.

So this function builds a partial App type, then uses it to run AppM actions to add the rest of the data before returning it.