Restyler.Config.Include

Documentation

data Include

Constructors

Include Pattern
**/*.hs

Negated Pattern
!**/*.temp

includePath :: [Include] -> FilePath -> Bool

Determine if a set of Includes match a file

Don't try to over-think this. It works how you would expect, and you can confirm in its test cases.