Restyler.Model.Config.ExpectedKeys

Documentation

validateObjectKeys :: [String] -> HashMap Text v -> Parser ()

Validate there are no unexpected keys in an Object

This is provided for convenience in the most common use-case. For a more flexible interface, see validateExpectedKeyBy.

validateExpectedKeyBy

Arguments

:: (Eq c, Show c)

=> String
The label to show as Unknown <label> ...

-> (a -> [c])
A function to project each valid value as a comparable key

-> [a]
The input list of valid items

-> [c]
The input key

-> Either String a

Validate that a key is present in a list of (projected) items

Returns the item found when validation passes.

N.B. the [c] values are almost always going to be String. It's typed as a list of c because the only actual requirement is that there are equatable elements so we can compute edit-distances.

>>> validateExpectedKeyBy "key" fst [("foo", 1), ("bar", 2)] ("foo" :: String)
Right ("foo",1)