Restyler.Model.PullRequest.Status

Documentation

data PullRequestStatus

Constructors

NoDifferencesStatus
We found no differences after restyling

DifferencesStatus URL
We found differences and opened a restyled PullRequest

ErrorStatus URL
We encountered an error and can link to a Job

sendPullRequestStatus :: MonadApp m => PullRequestStatus -> m ()

Send a PullRequestStatus for the original Pull Request

sendPullRequestStatus_ :: MonadApp m => PullRequestStatus -> m ()

sendPullRequestStatus but ignore any exceptions

This is useful for emitting the Errored status, where we wouldn't want an exception here to muddy the debugging of the error we're reporting.