Restyler.Model.Restyler

Documentation

data Restyler

How to run a given restyler

Constructors

Restyler

Fields

rName :: String

Unique name for this restyler, not configurable

rImage :: String

Docker image for this restyler, not configurable

rCommand :: [String]

Command to run, usually the name, not configurable

If there are arguments that are always required, such as a sub-command or an inplace flag, they should be present here. This is to avoid users accidentally overriding them out and being confused.

rArguments :: [String]

Arguments to pass before the paths

rInclude :: [Include]

Patterns to match for files to restyle

rInterpreters :: [Interpreter]

Interpreters to check for

rSupportsArgSep :: Bool

Can we pass -- between arguments and paths

rSupportsMultiplePaths :: Bool

Can we pass multiple paths at once

defaultRestylers :: [Restyler]

allRestylers :: [Restyler]

namedRestyler :: MonadPlus m => Text -> m Restyler

unsafeNamedRestyler :: Text -> Restyler